+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 - Show failed step error details in the summary formatter output
 - Fixed up JRuby examples which weren't running due to anglicisation issues (Pivoted to use English step definitions to help JRuby testing)
 - Fixed up Arabic example which had some incorrect logic for step definition matching (Due to RTL nature of the language)
+- Disable ANSI color by default for console formatter output written with `--out`.
 
 ## [11.1.0] - 2026-06-02
 ### Added